Move AppWorks component to matters dir
