+++ pages/index.vue
@@ -90,7 +90,7 @@ import AppProfileIcon from '~/components/matters/AppProfileIcon'
 import AppSkillButton from '~/components/matters/AppSkillButton'
 import AppText from '~/components/elements/AppText'
 import AppWishListButton from '~/components/matters/AppWishListButton'
-import AppWorks from '~/components/AppWorks'
+import AppWorks from '~/components/matters/AppWorks'
 
 export default {
   components: {
